wanone-backend/requirements.txt
flask==2.0.1
flask-restx==0.5.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
werkzeug==2.0.1
python-dotenv==0.19.0
pymysql==1.0.2
transformers==4.30.2
peft==0.4.0
torch
sqlalchemy==1.4.22